backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
paramiko==3.4.0
anthropic==0.7.8
python-socketio==5.10.0
aiofiles==23.2.1
python-dotenv==1.0.0
slowapi==0.1.9
markdown==3.5.1
weasyprint==60.1
markdown2==2.4.10
xhtml2pdf==0.2.11

